Ensure worker registers health and metrics endpoints.
