--- lsif/src/server/routes/meta.ts
@@ -1,15 +1,0 @@
-import express from 'express'
-import promClient from 'prom-client'
-
-/** Create a router containing health and metrics endpoint. */
-export function createMetaRouter(): express.Router {
-    const router = express.Router()
-    router.get('/ping', (_, res) => res.send('ok'))
-    router.get('/healthz', (_, res) => res.send('ok'))
-    router.get('/metrics', (_, res) => {
-        res.writeHead(200, { 'Content-Type': 'text/plain' })
-        res.end(promClient.register.metrics())
-    })
-
-    return router
-}
