+++ lsif/src/shared/api/init.ts
@@ -7,17 +7,19 @@ import { makeMetricsMiddleware } from './middleware/metrics'
 import { Tracer } from 'opentracing'
 import { Logger } from 'winston'
 
-export function makeExpressApp({
+export function startExpressApp({
     routes,
+    port,
     logger,
     tracer,
     selectHistogram = () => undefined,
 }: {
     routes: express.Router[]
+    port: number
     logger: Logger
     tracer?: Tracer
     selectHistogram?: (route: string) => promClient.Histogram<string> | undefined
-}): express.Express {
+}): void {
     const loggingOptions = {
         winstonInstance: logger,
         level: 'debug',
@@ -30,6 +32,7 @@ export function makeExpressApp({
     app.use(tracingMiddleware({ tracer }))
     app.use(loggingMiddleware(loggingOptions))
     app.use(makeMetricsMiddleware(selectHistogram))
+    app.use(createMetaRouter())
 
     for (const route of routes) {
         app.use(route)
@@ -39,5 +42,18 @@ export function makeExpressApp({
     // will apply to all routes and other middleware.
     app.use(errorHandler(logger))
 
-    return app
+    app.listen(port, () => logger.debug('API server listening', { port }))
+}
+
+/** Create a router containing health and metrics endpoint. */
+function createMetaRouter(): express.Router {
+    const router = express.Router()
+    router.get('/ping', (_, res) => res.send('ok'))
+    router.get('/healthz', (_, res) => res.send('ok'))
+    router.get('/metrics', (_, res) => {
+        res.writeHead(200, { 'Content-Type': 'text/plain' })
+        res.end(promClient.register.metrics())
+    })
+
+    return router
 }
